added @PreDestroy methods, config of threadpool now in external file
added @PreDestroy methods to properly shutdown the application (shutdown threadpool + kill all running processes of SunsetExecutor).
moved configuration of threadpool to external *.properties file to make it independent of the generated *.jar-file; changed SunsetThreadPoolConfiguration class accordingly.
updated *.bat file to shutdown the springboot application and all of its subprocesses (works, but it is not a graceful shutdown).
added Error405 to CustomErrorController.